requirements.txt
github3.py
coala-bears[all-deps]
